feat: make slides work & components refactor Add css files for reveal to show in component. Refactor components files into their own directory
