+++ src/components/Header/Header.tsx
@@ -3,7 +3,7 @@ import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import Button from "@mui/material/Button";
-import { useUserProfile } from "./useUserProfile.ts";
+import { useUserProfile } from "../../hooks/useUserProfile.ts";
 
 const Header = () => {
     const user = useUserProfile();
