+++ src/components/Presentation/Presentation.tsx
@@ -1,26 +1,28 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
 import Reveal from 'reveal.js';
-import Markdown from 'reveal.js/plugin/markdown/markdown.esm.js';
+import '/node_modules/reveal.js/dist/reset.css'
 import '/node_modules/reveal.js/dist/reveal.css';
 import '/node_modules/reveal.js/dist/theme/black.css';
+import './presentation.css';
 import { useEffect } from "react";
 
 const Presentation = () => {
     useEffect(() => {
-        const deck = new Reveal({
-            plugins: [ Markdown ]
-        })
-        deck.initialize();
-    }, []);
+        const deck = new Reveal();
+        deck.initialize({
+            embedded: true,
+        });
+    });
 
     return (
-        <>
+        <div>
             <div className="reveal">
                 <div className="slides">
-                    <section>Slide 1</section>
-                    <section>Slide 2</section>
+                    <section>SLIDE 1</section>
                 </div>
             </div>
-        </>
+        </div>
     )
 };
