+++ src/components/Presentation/presentation.css
@@ -1,0 +1,5 @@
+.reveal {
+    position:absolute;
+    height: 70%;
+    width: 70%;
+}
